+++ packages/gatsby/src/redux/run-sift.js
@@ -275,7 +275,13 @@ function handleMany(siftArgs, nodes, sort) {
       .map(field => field.replace(/___/g, `.`))
       .map(field => v => _.get(v, field))
 
-    result = _.orderBy(result, convertedFields, sort.order)
+    // Gatsby's sort interface only allows one sort order (e.g `desc`)
+    // to be specified. However, multiple sort fields can be
+    // provided. This is inconsistent. The API should allow the
+    // setting of an order per field. Until the API can be changed
+    // (probably v3), we apply the sort order to the first field only,
+    // implying asc order for the remaining fields.
+    result = _.orderBy(result, convertedFields, [sort.order])
   }
   return result
 }