+++ packages/gatsby-plugin-react-helmet/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.0.4"></a>
+
+## [3.0.4](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.3...gatsby-plugin-react-helmet@3.0.4) (2018-12-01)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+<a name="3.0.3"></a>
+
+## [3.0.3](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.2...gatsby-plugin-react-helmet@3.0.3) (2018-11-29)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
+<a name="3.0.2"></a>
+
+## [3.0.2](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.1...gatsby-plugin-react-helmet@3.0.2) (2018-11-12)
+
+**Note:** Version bump only for package gatsby-plugin-react-helmet
+
 <a name="3.0.1"></a>
 
 ## [3.0.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-react-helmet/compare/gatsby-plugin-react-helmet@3.0.0...gatsby-plugin-react-helmet@3.0.1) (2018-10-29)